test: rename tests/framework/testutils to tests/framework/utils
Please read https://go.dev/doc/effective_go#package-names
