+++ tests/common/status_test.go
@@ -20,7 +20,7 @@ import (
 	"time"
 
 	"go.etcd.io/etcd/tests/v3/framework/config"
-	"go.etcd.io/etcd/tests/v3/framework/testutils"
+	testutils "go.etcd.io/etcd/tests/v3/framework/utils"
 )
 
 func TestStatus(t *testing.T) {
