+++ tests/e2e/etcd_grpcproxy_test.go
@@ -28,7 +28,7 @@ import (
 	"go.etcd.io/etcd/pkg/v3/expect"
 	"go.etcd.io/etcd/tests/v3/framework/config"
 	"go.etcd.io/etcd/tests/v3/framework/e2e"
-	"go.etcd.io/etcd/tests/v3/framework/testutils"
+	testutils "go.etcd.io/etcd/tests/v3/framework/utils"
 )
 
 func TestGrpcProxyAutoSync(t *testing.T) {
