+++ tests/framework/e2e/flags.go
@@ -19,7 +19,7 @@ import (
 	"os"
 	"runtime"
 
-	"go.etcd.io/etcd/tests/v3/framework/testutils"
+	testutils "go.etcd.io/etcd/tests/v3/framework/utils"
 )
 
 var (
